+++ app/src/main/java/com/infomaniak/swisstransfer/ui/screen/main/MainNavHost.kt
@@ -17,6 +17,7 @@
  */
 package com.infomaniak.swisstransfer.ui.screen.main
 
+import FilesDetailsScreen
 import androidx.compose.animation.EnterTransition
 import androidx.compose.animation.ExitTransition
 import androidx.compose.animation.fadeIn
@@ -32,7 +33,8 @@ import com.infomaniak.swisstransfer.ui.navigation.MainNavigation.*
 import com.infomaniak.swisstransfer.ui.navigation.MainNavigation.ReceivedDestination.Companion.receivedDestination
 import com.infomaniak.swisstransfer.ui.screen.main.settings.SettingsScreenWrapper
 import com.infomaniak.swisstransfer.ui.screen.main.transfers.TransfersScreenWrapper
-import com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles.FilesDetailsComponent
+import com.infomaniak.swisstransfer.ui.theme.LocalWindowAdaptiveInfo
+import com.infomaniak.swisstransfer.ui.utils.isWindowSmall
 
 @Composable
 fun MainNavHost(
@@ -72,7 +74,7 @@ fun MainNavHost(
         composable<SettingsDestination> { SettingsScreenWrapper() }
         composable<FilesDetailsDestination> {
             val filesDetailsDestination: FilesDetailsDestination = it.toRoute()
-            FilesDetailsComponent(
+            FilesDetailsScreen(
                 navigateToDetails = { folderUuid ->
                     navController.navigate(FilesDetailsDestination(folderUuid))
                 },
@@ -84,7 +86,7 @@ fun MainNavHost(
                 withFilesSize = false,
                 withSpaceLeft = false,
                 withFileDelete = false,
-                shouldDisplayTopAppBar = isWindowSmall,
+                shouldDisplayTopAppBar = LocalWindowAdaptiveInfo.current.isWindowSmall(),
             )
         }
     }
