+++ app/src/main/java/com/infomaniak/swisstransfer/ui/screen/newtransfer/NewTransferNavHost.kt
@@ -26,9 +26,10 @@ import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
 import com.infomaniak.swisstransfer.ui.navigation.NewTransferNavigation
 import com.infomaniak.swisstransfer.ui.navigation.NewTransferNavigation.*
-import com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles.FilesDetailsScreen
 import com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles.ImportFilesScreen
 import com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles.components.TransferTypeUi
+import com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles.NewTransferFilesDetailsScreen
+import com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles.ValidateUserEmailScreen
 import com.infomaniak.swisstransfer.ui.screen.newtransfer.upload.UploadErrorScreen
 import com.infomaniak.swisstransfer.ui.screen.newtransfer.upload.UploadProgressScreen
 import com.infomaniak.swisstransfer.ui.screen.newtransfer.upload.UploadSuccessScreen
@@ -55,8 +56,8 @@ fun NewTransferNavHost(
                     navController.navigate(ValidateUserEmailDestination(email, recipients))
 				},
                 navigateToFilesDetails = {
-                    navController.navigate(FilesDetailsDestination)
-                },
+                    navController.navigate(NewTransferFilesDetailsDestination)
+                }
             )
         }
         composable<ValidateUserEmailDestination> {
@@ -117,11 +118,11 @@ fun NewTransferNavHost(
                 navigateBackToImportFiles = {
                     navController.popBackStack(route = ImportFilesDestination, inclusive = false)
                 },
-        composable<FilesDetailsDestination> {
+        composable<NewTransferFilesDetailsDestination> {
             val backStackEntry = remember(it) { navController.getBackStackEntry(ImportFilesDestination) }
-            FilesDetailsScreen(
+            NewTransferFilesDetailsScreen(
                 importFilesViewModel = hiltViewModel<ImportFilesViewModel>(backStackEntry),
-                navigateToDetails = { _ -> navController.navigate(FilesDetailsDestination) },
+                navigateToDetails = { _ -> navController.navigate(NewTransferFilesDetailsDestination) },
                 withFilesSize = true,
                 withSpaceLeft = true,
                 withFileDelete = true,
