+++ app/src/main/java/com/infomaniak/swisstransfer/ui/screen/newtransfer/importfiles/NewTransferFilesDetailsScreen.kt
@@ -17,32 +17,26 @@
  */
 package com.infomaniak.swisstransfer.ui.screen.newtransfer.importfiles
 
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.padding
+import FilesDetailsScreen
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.PreviewParameter
-import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.infomaniak.multiplatform_swisstransfer.common.interfaces.ui.FileUi
-import com.infomaniak.swisstransfer.ui.components.FileItemList
 import com.infomaniak.swisstransfer.ui.components.SwissTransferTopAppBar
 import com.infomaniak.swisstransfer.ui.components.TopAppBarButton
 import com.infomaniak.swisstransfer.ui.previewparameter.FileUiListPreviewParameter
-import com.infomaniak.swisstransfer.ui.screen.newtransfer.FilesSize
 import com.infomaniak.swisstransfer.ui.screen.newtransfer.ImportFilesViewModel
-import com.infomaniak.swisstransfer.ui.theme.Margin
 import com.infomaniak.swisstransfer.ui.theme.SwissTransferTheme
 import com.infomaniak.swisstransfer.ui.utils.PreviewAllWindows
+import getOnFileRemoveCallback
 
 @Composable
-fun FilesDetailsScreen(
+fun NewTransferFilesDetailsScreen(
     importFilesViewModel: ImportFilesViewModel = hiltViewModel<ImportFilesViewModel>(),
     folderUuid: String? = null,
     navigateToDetails: (String) -> Unit,
@@ -61,7 +55,7 @@ fun FilesDetailsScreen(
     }
 
     files?.let {
-        FilesDetailsScreen(
+        NewTransferFilesDetailsScreen(
             files = it,
             navigateToDetails = navigateToDetails,
             withFilesSize = withFilesSize,
@@ -72,17 +66,8 @@ fun FilesDetailsScreen(
     }
 }
 
-private fun getOnFileRemoveCallback(
-    importFilesViewModel: ImportFilesViewModel,
-    withFileDelete: Boolean,
-): ((String) -> Unit)? {
-    return if (withFileDelete) {
-        { importFilesViewModel.removeFileByUid(it) }
-    } else null
-}
-
 @Composable
-private fun FilesDetailsScreen(
+private fun NewTransferFilesDetailsScreen(
     files: List<FileUi>,
     navigateToDetails: (String) -> Unit,
     withFilesSize: Boolean,
@@ -95,7 +80,7 @@ private fun FilesDetailsScreen(
             SwissTransferTopAppBar(navigationMenu = TopAppBarButton.backButton { navigateBack() })
         }
     ) { paddingValues ->
-        FilesDetailsComponent(
+        FilesDetailsScreen(
             paddingValues = paddingValues,
             files = files,
             navigateToDetails = navigateToDetails,
@@ -106,81 +91,12 @@ private fun FilesDetailsScreen(
     }
 }
 
-@Composable
-fun FilesDetailsComponent(
-    shouldDisplayTopAppBar: Boolean,
-    paddingValues: PaddingValues = PaddingValues(0.dp),
-    importFilesViewModel: ImportFilesViewModel = hiltViewModel<ImportFilesViewModel>(),
-    folderUuid: String? = null,
-    navigateToDetails: (String) -> Unit,
-    withFilesSize: Boolean,
-    withSpaceLeft: Boolean,
-    withFileDelete: Boolean,
-    navigateBack: (() -> Unit),
-    close: (() -> Unit),
-) {
-    // If we don't have a folderUuid, it means we have to load files from importedFiles in ImportFilesViewModel
-    val files by importFilesViewModel.files.collectAsStateWithLifecycle()
-
-    if (files?.isEmpty() == true) navigateBack()
-
-    LaunchedEffect(Unit) {
-        importFilesViewModel.loadFiles(folderUuid)
-    }
-
-    files?.let {
-        Column {
-            if (shouldDisplayTopAppBar) {
-                SwissTransferTopAppBar(
-                    navigationMenu = TopAppBarButton.backButton { navigateBack() },
-                    actionMenus = arrayOf(TopAppBarButton.closeButton { close() }),
-                )
-            }
-
-            FilesDetailsComponent(
-                paddingValues = paddingValues,
-                files = it,
-                navigateToDetails = navigateToDetails,
-                withFileSize = withFilesSize,
-                withSpaceLeft = withSpaceLeft,
-                onFileRemoved = getOnFileRemoveCallback(importFilesViewModel, withFileDelete),
-            )
-        }
-    }
-}
-
-@Composable
-fun FilesDetailsComponent(
-    paddingValues: PaddingValues,
-    files: List<FileUi>,
-    navigateToDetails: (String) -> Unit,
-    withFileSize: Boolean,
-    withSpaceLeft: Boolean,
-    onFileRemoved: ((uuid: String) -> Unit)? = null,
-) {
-    Column(modifier = Modifier.padding(paddingValues)) {
-        FilesSize(files, withFilesSize = withFileSize, withSpaceLeft)
-        FileItemList(
-            modifier = Modifier.padding(horizontal = Margin.Medium),
-            files = files,
-            isRemoveButtonVisible = onFileRemoved != null,
-            isCheckboxVisible = { false },
-            isUidChecked = { false },
-            setUidCheckStatus = { _, _ -> },
-            onRemoveUid = { onFileRemoved?.invoke(it) },
-            onClick = {
-                navigateToDetails(it)
-            }
-        )
-    }
-}
-
 @PreviewAllWindows
 @Composable
-private fun FilesDetailsScreenPreview(@PreviewParameter(FileUiListPreviewParameter::class) files: List<FileUi>) {
+private fun Preview(@PreviewParameter(FileUiListPreviewParameter::class) files: List<FileUi>) {
     SwissTransferTheme {
         Surface {
-            FilesDetailsScreen(
+            NewTransferFilesDetailsScreen(
                 files = files,
                 navigateToDetails = { _ -> },
                 withFilesSize = true,
